+++ Baganator/CategoryViews/BagLayout.lua
@@ -107,9 +107,9 @@ end
 addonTable.CategoryViews.BagLayoutMixin = {}
 
 function addonTable.CategoryViews.BagLayoutMixin:OnLoad()
-  self.labelsPool = CreateFramePool("Button", self:GetParent().Container, "BaganatorCategoryViewsCategoryButtonTemplate")
+  self.labelsPool = (BAGANATOR_335 and addonTable.Compatibility.CreateFramePool or CreateFramePool)("Button", self:GetParent().Container, "BaganatorCategoryViewsCategoryButtonTemplate")
   self.sectionButtonPool = addonTable.CategoryViews.GetSectionButtonPool(self:GetParent().Container)
-  self.dividerPool = CreateFramePool("Frame", self:GetParent().Container, "BaganatorBagDividerTemplate")
+  self.dividerPool = (BAGANATOR_335 and addonTable.Compatibility.CreateFramePool or CreateFramePool)("Frame", self:GetParent().Container, "BaganatorBagDividerTemplate")
 
   self.notShown = {}